refactor: replace toString with toFixed
