+++ packages/crypto/lib/models/block.js
@@ -1,7 +1,7 @@
 const { cloneDeepWith } = require('lodash')
 const { createHash } = require('crypto')
 const ByteBuffer = require('bytebuffer')
-const Bignum = require('../utils/bignum')
+const { Bignum, bignumify } = require('../utils')
 const Transaction = require('./transaction')
 const configManager = require('../managers/config')
 const { crypto, slots } = require('../crypto')
@@ -162,7 +162,7 @@ module.exports = class Block {
 
   static getId (data) {
     const idHex = Block.getIdHex(data)
-    return new Bignum(idHex, 16).toString()
+    return new Bignum(idHex, 16).toFixed()
   }
 
   /**
@@ -337,7 +337,7 @@ module.exports = class Block {
     block.timestamp = buf.readUInt32(4)
     block.height = buf.readUInt32(8)
     block.previousBlockHex = buf.slice(12, 20).toString('hex')
-    block.previousBlock = new Bignum(block.previousBlockHex, 16).toString()
+    block.previousBlock = new Bignum(block.previousBlockHex, 16).toFixed()
     block.numberOfTransactions = buf.readUInt32(20)
     block.totalAmount = new Bignum(buf.readUInt64(24))
     block.totalFee = new Bignum(buf.readUInt64(32))
@@ -394,6 +394,8 @@ module.exports = class Block {
    * @static
    */
   static serialize (block, includeSignature = true) {
+    bignumify(block, ['totalAmount', 'totalFee', 'reward'])
+
     block.previousBlockHex = toBytesHex(block.previousBlock)
 
     const bb = new ByteBuffer(256, true)
@@ -402,9 +404,9 @@ module.exports = class Block {
     bb.writeUInt32(block.height)
     bb.append(block.previousBlockHex, 'hex')
     bb.writeUInt32(block.numberOfTransactions)
-    bb.writeUInt64(+block.totalAmount.toString())
-    bb.writeUInt64(+block.totalFee.toString())
-    bb.writeUInt64(+block.reward.toString())
+    bb.writeUInt64(+block.totalAmount.toFixed())
+    bb.writeUInt64(+block.totalFee.toFixed())
+    bb.writeUInt64(+block.reward.toFixed())
     bb.writeUInt32(block.payloadLength)
     bb.append(block.payloadHash, 'hex')
     bb.append(block.generatorPublicKey, 'hex')
