+++ packages/crypto/lib/utils/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = {
+  Bignum: require('./bignum'),
+  bignumify: require('./bignumify'),
+  sortTransactions: require('./sort-transactions')
+}
